GNN-RAG inference pipeline added. Some renames.
